+++ library/decorators/__init__.py
@@ -1,1 +1,2 @@
 from .is_superuser import is_superuser
+from .id_is_real import id_is_real
