+++ templates/404_page.html
@@ -45,7 +45,7 @@
                                 </div>
                             {% endif %}
                             <div class="row justify-content-center">
-                                <a class="btn btn-outline-light" href="{% url 'index' %}">HOME</a>
+                                <a class="btn btn-outline-light" href="{{ previous_url }}">برگشت</a>
                             </div>
                         </div>
                     </div>
