+++ vidoneplus/views/__init__.py
@@ -1,0 +1,3 @@
+from .index import index
+from .page_not_found import page_not_found
+from .validate_course import validate_course
