+++ vidoneplus/views/index.py
@@ -1,0 +1,5 @@
+from .imports import *
+
+
+def index(request):
+    return render(request, 'client/dashboard.html')
